requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
readability-lxml==0.8.1
psycopg2-binary==2.9.9
